+++ packages/app/src/sections/exchange/BasicSwap.tsx
@@ -3,6 +3,7 @@ import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
 import * as Text from 'components/Text'
+import { FullHeightContainer, MainContent } from 'styles/common'
 
 import SwapCurrencies from './SwapCurrencies'
 import BaseCurrencyCard from './TradeCard/Cards/BaseCurrencyCard'
@@ -13,21 +14,27 @@ const BasicSwap: FC = memo(() => {
 	const { t } = useTranslation()
 
 	return (
-		<>
-			<ExchangeTitle>{t('exchange.synth-exchange')}</ExchangeTitle>
-			<PageWidthContainer>
-				<DesktopCardsContainer>
-					<QuoteCurrencyCard />
-					<SwapCurrencies />
-					<BaseCurrencyCard />
-				</DesktopCardsContainer>
-				<FooterCard />
-			</PageWidthContainer>
-		</>
+		<StyledFullHeightContainer>
+			<MainContent>
+				<ExchangeTitle>{t('exchange.synth-exchange')}</ExchangeTitle>
+				<PageWidthContainer>
+					<DesktopCardsContainer>
+						<QuoteCurrencyCard />
+						<SwapCurrencies />
+						<BaseCurrencyCard />
+					</DesktopCardsContainer>
+					<FooterCard />
+				</PageWidthContainer>
+			</MainContent>
+		</StyledFullHeightContainer>
 	)
 })
 
 export default BasicSwap
+
+const StyledFullHeightContainer = styled(FullHeightContainer)`
+	padding-top: 14px;
+`
 
 const ExchangeTitle = styled(Text.Body).attrs({ weight: 'bold' })`
 	color: ${(props) => props.theme.colors.selectedTheme.button.text.primary};
@@ -39,9 +46,9 @@ const ExchangeTitle = styled(Text.Body).attrs({ weight: 'bold' })`
 const DesktopCardsContainer = styled.div`
 	display: flex;
 	flex-direction: column;
-	background: ${(props) => props.theme.colors.selectedTheme.cell.fill};
+	background: ${(props) => props.theme.colors.selectedTheme.newTheme.exchange.card};
 	border-radius: 10px;
-	border: ${(props) => props.theme.colors.selectedTheme.border};
+	border: ${(props) => props.theme.colors.selectedTheme.newTheme.border.style};
 	box-sizing: border-box;
 	position: relative;
 `
