+++ packages/app/src/sections/exchange/ExchangeContent.tsx
@@ -1,20 +1,15 @@
 import { memo } from 'react'
-import styled from 'styled-components'
 
 import { DesktopOnlyView, MobileOrTabletView } from 'components/Media'
 import BasicSwap from 'sections/exchange/BasicSwap'
 import ExchangeModals from 'sections/exchange/ExchangeModals'
 import { MobileSwap } from 'sections/exchange/MobileSwap'
-import { PageContent, FullHeightContainer, MainContent } from 'styles/common'
+import { PageContent } from 'styles/common'
 
 const ExchangeContent = memo(() => (
 	<PageContent>
 		<DesktopOnlyView>
-			<StyledFullHeightContainer>
-				<MainContent>
-					<BasicSwap />
-				</MainContent>
-			</StyledFullHeightContainer>
+			<BasicSwap />
 		</DesktopOnlyView>
 		<MobileOrTabletView>
 			<MobileSwap />
@@ -22,9 +17,5 @@ const ExchangeContent = memo(() => (
 		<ExchangeModals />
 	</PageContent>
 ))
-
-const StyledFullHeightContainer = styled(FullHeightContainer)`
-	padding-top: 14px;
-`
 
 export default ExchangeContent
